server/requirements.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
cloudinary==1.29.0
Django==3.2.7
django-cors-headers==3.13.0
django-dotenv==1.4.2
django-grappelli==3.0.3
django-mailjet==0.3.1
djangorestframework==3.13.1
idna==3.3
mailjet-rest==1.3.4
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stripe==3.5.0
urllib3==1.26.12
